+++ src/combinations/useFunc.ts
