+++ src/patternOperations/patternCount.ts
