+++ src/patternOperations/patternIntegration.ts
