+++ src/patternOperations/patternMatch.ts
