+++ src/patternOperations/patternUtils.ts
