+++ src/researchv7.ts
@@ -1,6 +1,6 @@
 import { getSubDir } from "./utils/getSubDir";
 import { getAllFiles } from "./utils/getAllFiles";
-import { useFunc1_v7, useFunc } from "./utils/useFunc";
+import { useFunc1_v7, useFunc } from "./combinations/useFunc";
 import { v7Idetifiers, v7removeException, deepImport } from "./utils/v7Idetifiers";
 import fs from 'fs';
 import path from 'path';
